+++ osidb/tests/models.py
@@ -1,7 +1,10 @@
 import uuid
 
+from django.core.cache import caches
 from django.db import models
+from rest_framework.throttling import UserRateThrottle
 
+from config.settings import DEBUG
 from osidb.mixins import AlertMixin
 from osidb.models import ComparableTextChoices
 
@@ -26,3 +29,12 @@ class ComparableTextChoices_1(ComparableTextChoices):
 
 class ComparableTextChoices_2(ComparableTextChoices):
     TEST = "TEST"
+
+
+class LowRateThrottle(UserRateThrottle):
+    """Throttle class with very low rate to test that throttling works."""
+
+    rate = "2/day"
+    # In the test env there is no cache but for throttling to work a cache is required
+    if DEBUG:
+        cache = caches["locmem"]